+++ packages/core/src/assembly/assembly.spec.ts
