+++ packages/core/src/assembly/random/lock.spec.ts
@@ -1,4 +1,3 @@
-import { LockedParts } from '#core/assembly/random/lock'
 
 import { boosterNotEquipped } from '@ac6_assemble_tool/parts/not-equipped'
 import { tank } from '@ac6_assemble_tool/parts/types/base/category'
@@ -7,6 +6,8 @@ import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
 import { random } from '@ac6_assemble_tool/shared/array'
 import { it, fc, test } from '@fast-check/vitest'
 import { describe, expect } from 'vitest'
+
+import { LockedParts } from './lock'
 
 import {
   genAssembly,
