+++ packages/core/src/assembly/random/random-assembly.spec.ts
@@ -1,14 +1,14 @@
+import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
+import { fc, it } from '@fast-check/vitest'
+import { afterEach, beforeEach, describe, expect, type Mock, vi } from 'vitest'
+
 import {
   OverTryLimitError,
   OverwriteInnerSecretValidatorError,
   RandomAssembly,
-} from '#core/assembly/random/random-assembly'
-import type { Validator } from '#core/assembly/random/validator/base'
-import { failure, success } from '#core/assembly/random/validator/result'
-
-import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
-import { fc, it } from '@fast-check/vitest'
-import { afterEach, beforeEach, describe, expect, type Mock, vi } from 'vitest'
+} from './random-assembly'
+import type { Validator } from './validator/base'
+import { failure, success } from './validator/result'
 
 describe(RandomAssembly.name, () => {
   it.prop([fc.array(generateValidatorWithKey())])(
