+++ packages/core/src/assembly/random/validator/validators.spec.ts
@@ -1,11 +1,3 @@
-import {
-  disallowArmsLoadOver,
-  disallowLoadOver,
-  notCarrySameUnitInSameSide,
-  notOverEnergyOutput,
-  totalCoamNotOverMax,
-  totalLoadNotOverMax,
-} from '#core/assembly/random/validator/validators'
 
 import type { LeftArmUnit, ArmUnit } from '@ac6_assemble_tool/parts/arm-units'
 import {
@@ -18,6 +10,15 @@ import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
 import { fc, it as fcit } from '@fast-check/vitest'
 import sinon from 'sinon'
 import { afterEach, beforeEach, describe, expect } from 'vitest'
+
+import {
+  disallowArmsLoadOver,
+  disallowLoadOver,
+  notCarrySameUnitInSameSide,
+  notOverEnergyOutput,
+  totalCoamNotOverMax,
+  totalLoadNotOverMax,
+} from './validators'
 
 import { genAssembly } from '#spec-helper/property-generator'
