+++ packages/core/src/assembly/serialize/as-query.spec.ts
@@ -1,11 +1,12 @@
-import {
-  assemblyToSearch,
-  searchToAssembly,
-} from '#core/assembly/serialize/as-query'
 
 import { it } from '@fast-check/vitest'
 import fc from 'fast-check'
 import { describe, expect } from 'vitest'
+
+import {
+  assemblyToSearch,
+  searchToAssembly,
+} from './as-query'
 
 import {
   genAssembly,
