+++ packages/core/src/assembly/store/repository/data-transfer-object.spec.ts
@@ -1,6 +1,6 @@
-import { storedAssemblyDtoScheme } from '#core/assembly/store/repository/data-transfer-object'
-
 import { describe, it, expect } from 'vitest'
+
+import { storedAssemblyDtoScheme } from './data-transfer-object'
 
 describe('StoredAssemblyDto スキーマ', () => {
   describe('storedAssemblyDtoScheme', () => {
