+++ packages/core/src/assembly/store/repository/indexed-db/stored-assembly.spec.ts
@@ -1,14 +1,16 @@
 import { createAssembly } from '#core/assembly/assembly'
-import { IndexedDbRepository } from '#core/assembly/store/repository/indexed-db/indexed-db-repository'
-import type {
-  ClearableStoredAssemblyRepository,
-  StoredAssemblyRepository,
-} from '#core/assembly/store/stored-assembly'
 
 import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
 import { it } from '@fast-check/vitest'
 import { ulid } from 'ulid'
 import { describe, expect } from 'vitest'
+
+import type {
+  ClearableStoredAssemblyRepository,
+  StoredAssemblyRepository,
+} from '../../stored-assembly'
+
+import { IndexedDbRepository } from './indexed-db-repository'
 
 import { genAssembly } from '#spec-helper/property-generator'
